backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-JWT-Extended>=4.0
Flask-Cors>=3.0
python-dotenv>=0.19
pymysql>=1.0
Flask-Migrate>=3.1
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
requests>=2.28.0
Jinja2>=3.1.0
bleach>=6.0.0
